+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Improvements
 
 - Made `IntId::is_sgi` public.
+- Made `IntId::is_*` methods const.
 
 ## 0.2.2